+++ src/components/Button.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { cva, type VariantProps } from "class-variance-authority"
 
-import { cn } from "@/components/ui/lib/utils"
+import { cn } from "@/lib/utils"
 
 const buttonVariants = cva("inline-flex items-center justify-center gap-2 whitespace-nowrap", {
   variants: {
