+++ src/components/ui/drawer.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { Drawer as DrawerPrimitive } from "vaul"
 
-import { cn } from "@/components/ui/lib/utils"
+import { cn } from "@/lib/utils"
 
 const Drawer = ({
   shouldScaleBackground = true,
