+++ src/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/components/ui/lib/utils"
+import { cn } from "@/lib/utils"
 
 function Skeleton({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
   return <div className={cn("animate-pulse rounded-md bg-primary/10", className)} {...props} />
