+++ src/components/ui/switch.tsx
@@ -1,7 +1,7 @@
 import * as SwitchPrimitives from "@radix-ui/react-switch"
 import * as React from "react"
 
-import { cn } from "@components/ui/lib/utils"
+import { cn } from "@/lib/utils"
 
 const Switch = React.forwardRef<
   React.ElementRef<typeof SwitchPrimitives.Root>,
