+++ src/pages/orders/active/components/OrderMenuItem.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/components/ui/lib/utils"
+import { cn } from "@/lib/utils"
 import React from "react"
 
 interface MenuItem {
